feat/refactor: split up into two modules
