+++ Cargo.toml
@@ -1,14 +1,7 @@
-[package]
-name = "grrs"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+members = [
+    "services/cli",
+]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ab_glyph = "0.2.28"
-anyhow = "1.0.86"
-clap = { version = "4.0", features = ["derive"] }
-env_logger = "0.11.3"
-imageproc = "0.25.0"
-log = "0.4.22"
+[workspace.dependencies]
+renderer = { path = "libs/renderer" }
