+++ README.md
@@ -1,5 +1,13 @@
 # glasses-of-mishra
 
+Example CLI usage:
+
 ```sh
-cargo run -- 'Hello, World!'
+cargo run -- -c='Hello, world!' -t=templates/blue.png -o=hello.png
+```
+
+Get help:
+
+```sh
+cargo run -- -h
 ```
