+++ libs/renderer/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "renderer"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ab_glyph = "0.2.28"
+anyhow = "1.0.86"
+imageproc = "0.25.0"
