+++ services/cli/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.86"
+clap = { version = "4.0", features = ["derive"] }
+renderer = { workspace = true }
